+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ### Added
 - Typescript definitions
 - Watch the included files in static templates.
-- AppVeyor Test.
+- AppVeyor tests.
 
 ### Changed
 - peerDependencies has rollup>=0.61 to allow dependency detection (see Rollup [#2259](https://github.com/rollup/rollup/pull/2259))